Altered the structure to split key portions of functionality to their respective packages and areas
